Feat: 장소 메뉴 조회 request 타입 변경

+++ data/src/main/kotlin/com/acon/data/api/remote/SpotApi.kt
@@ -23,6 +23,6 @@ interface SpotApi {
 
     @GET("/api/v1/spot/{spotId}/menus")
     suspend fun getSpotMenuList(
-        @Path ("spotId") spotId: Int
+        @Path ("spotId") spotId: Long
     ): SpotDetailMenuListResponse
 }
